Move rejection reasons from locale to partials
Working with a translation entry and interpolating values was about to
get a bunch more complicated because we are about to introduce an
interpoliation for student loans that doesn't exist for maths and
physics: the financial year that is being claimed for. Moving these
reasons to partials makes it easier to have policy-specific content in
each set of rejection reasons.