Allow service operators to navigate to decision page
Currently if a service operator exits the journey on the last check they
are unable to navigate back to the decision page.

So that service operators can get back to the decision page at any time
we now include a link to the decision on the index of tasks
along with the result of the decision if there is one.

I've extracted out `govuk-tag--warning` and `govuk-tag--alert` as these
supplement the GOV.UK tag component and reused the classes in our
custom tag component styles.

The link to "Approve or reject this claim" exists on claims that already
have a decision. If a user clicks this link on decided claims they will
see a notice saying "Claim outcome already decided" and redirects to
the claim#show page. This isn't a great user experience but it's a bit
of an edge case in the current system, there's no obvious user journey
to get to this point. In the future we may want to introduce a
decisions#show page to display information about the decision.