Links that look like buttons now work if JS
We can add a role and a data-module="govuk-button" to links that are
styled as buttons. This allows JS to pickup if space is pressed when
it has focus.

Whilst this will still me that non-JS users can not trigger these link
buttons when pressing space, this is still a big improvement so we
should probably ship it and think about if we want to invest the time
to change them to be proper buttons.