Fix Slack notifications
Previously, both Slack notification steps were wrapped in one job, which had
a `dependsOn` of the previous job. This meant that the job would only run
if `clean_up` had run successfully. Moving the failure message to its own job
with a `condition` of `failed()` means this job only runs if any of the previous
jobs have failed.