Warn about incomplete checks on decision screen
If a claim has not been fully checked, then we want to warn the service
operator about this when they come to make a decision.

To implement the error summary from the prototype, which represents
errors that don’t prevent the claim from being approved, I customised
the govuk-error-summary and govuk-error-summary__list components
following the Design System guidance [1]. To colour the links and
border, I copied and modified the styling from _error-summary.scss in
govuk-frontend.

[1] https://design-system.service.gov.uk/get-started/extending-and-modifying-components/#small-modifications-to-components