Deploy to development in pipelines
This uses the Azure Pipeline to deploy as well as run the tests if the branch
is master. This keeps the two things tightly coupled, and also allows us to
keep our deployment steps tightly coupled to our code, rather than having
them defined in Azure DevOps via a UI. This can also be adapted to deploy
to production when we're ready to do so.

I've also put all our Docker-related tasks in a seperate bin script, so it's easier
to follow, rather than having to parse a massive bunch of YAML.