MatchingAttributeFinder searches by policy
We've had reports of a number of false positives when searching for
matching attributes, because a number of applicants are eligible for both
policies.

This adds the source claim's policy to the `claims_to_compare` query,
ensuring that only claims for the source claim's policy are returned when
searching for potential duplicate claims.

I've also removed the optional `claims_to_compare` parameter as it's not
used anywhere, and complicates the class somewhat without any real
benefit.