Add Claim#incomplete_check_names method
I wanted to create a place for this logic to live. Even though it’s
simple logic right now, it’s going to get more complicated as we start
to make the list of checks be dynamic based on policy and claim.

It’s a bit mucky that the model is referring to a constant defined in
the controller, but I’m pretty confident that the constant will be moved
elsewhere in the near future so not too bothered by it.