ClaimCheckingTask to model checking tasks
We're about to introduce tasks that only apply to certain policies
(student loan amount) and tasks that will only appear for certain claims
(claims with matching details and identity confirmation). This new class
will allow us to model that and dynamically return the tasks for a given
claim based on its type and characteristics.