Admins can upload the DQT report CSV
Some of the headers in the CSV don't make much sense (ie. text1/2) but
hopefully their usage is clear, we considered adding a header converter
to give them more meaningful names but it presented a problem when
playing back error messages in the UI to the user.

It's possible for nil values to be parsed from the CSV where there
is no record in DQT.

Co-authored-by: Lorna Harwood <lorna@dxw.com>