Add a 'claims with matching details' task
This now checks if any submitted claims have values that match the source
claim. If there are, then a new task is added for the service operator to view
what these matching attributes are, allowinfg then to verify that these won't
cause any problems, or are potentially fraudulent before moving on to the
next task.