Use `concat` to build SERVICE_ENTITY_IDS array
Running `string` against a created array now appears to flatten out the array into a comma seperated list, so we need to change how we build the
`SERVICE_ENTITY_IDS` JSON array. This builds up the array using a
combination of a `[`, an (escaped) single quote `''`, the VSP entity ID and a
`]`. The means if the value of `VSP.VERIFY_ENTITY_ID` was
`http://example.com`, the resulting JSON string would be
`['http://example.com']`