Run `prettier:fix`
The new version of Prettier has some changes including always adding a
space after the `function` keyword and some changes to how HTML classes
are handled. Trailing commas are also now required in objects and arrays, but
I've disabled this in the `.prettierrc` file, as this can effect IE8 compatibility.