Set healthcheck endpoint to monitor
This tells Azure to monitor the healthcheck endpoint every 2 minutes.
If an instance does not respond within 10 minutes, the instance is taken
out of the pool. It's a manual process to put the app back in the pool,
and we're not (currently) able to be alerted if something is taken out
of the pool (this is a Windows-only thing at the moment).

More on the Health Check functionality is [here](https://github.com/projectkudu/kudu/wiki/Health-Check-%28Preview%29)