Use Rails’ translate method over I18n.translate
Rails’ `translate` method delegates to `I18n#translate`, but also
performs additional functions, including raising an exception when a
translation key is missing and the application is configured with
`ActionView::Base.raise_on_missing_translations` set to true. By using
this method we make it less likely that we will ship code where
translations are missing as we should see exceptions raised in the
specs.