Use an example DQT report CSV fixture in the specs
We test the overall behaviour in the DQTReportConsumer model and the
feature spec. I've created two helpers to allow us to reuse the fixture
file, one to create claims that exist in the file and one to return
the file itself. This should give future developers a concrete example
of the shape of the file we expect from DQT and the scenarios in which
we will create an automated check.