Transform DQT report rows to unique records
When we built the automated DQT checks we were working on the assumption
that there would be a single row for each record. It turns out there
can be multiple rows for each claimant.

We need to munge these rows down to unique records so we can iterate
through them and safely create checks for each claim.

The claim reference, date of birth and QTS award year are all consistent
between grouped rows so we can use the first row's value. For the ITT
subject and degree JAC codes we need to collate and combine these values
from the multiple grouped rows.

As this new class is responsible for transforming the data in to a shape
we expect I used the opportunity to move the logic to do with
parsing the dates and transforming columns to arrays.