Add a runbook for first-line support developers
I identified a few tasks that I’ve needed to do frequently in these
situations:

- dxw helpdesk first-line support dev
- developer support rotation whilst working on this service

This document will also hopefully let us build an onboarding checklist
for someone in DfE who needs to set up accounts for new dxw support
developers.

This document is mainly aimed at developers in dxw who will be taking
part in the first-line support rotation for the project. It’s written in
a way that it could also be used by e.g. an internal DfE team who are
supporting the project. But, a developer at DfE might find that this
document takes its time to explain things that they already know well.