Rework request specs relating to authentication
Make these specs more focused on the behaviour under test, i.e. the
behaviour of Admin::AuthController and it's integration with DfE
Sign-in.

Unlike the other request specs that use the sign_in_to_admin_with_role
helper, I've made these explicitly stub the auth callbacks and follow
the redirect to the callback action to make it a little more obvious
what is going on.

Also makes a slight tweak to the sign-out behaviour to send the user to
the sign-in page to avoid an unnecessary double-redirect.