Ensure the surname in DQT matches the one in the claim
Currently claims checkers do an extra informal identity verification
on the record in DQT to make sure it matches the claimant. They check
the name and National Insurance number (NiNo). The NiNo doesn't always
exist in DQT so instead we can add the extra verification step as a
match against the surname.

We recognise that names are subject to change and can legitimately not
match, but these claims will still be checked just manually by a person
instead of automatically.