Improve inconsistent payment details message
This splits the inconsistent payment details message into two paragraphs.
The first explains the problem, the second what the service operator can
do to resolve the problem.

There was some debate about putting in a link to the service operator
playbook but we were not overly happy about putting html in the
translation file or adding in a link that may change.