Move inconsistent payroll content inline
We've over-egged it a bit with static content in the locale files.
The extra indirection makes it harder for developers and content folks
to see the content in context and make changes to it. If the content is
(mostly) straight-forward static content shown in one place, let's
default to inlining it.

I've also loosened the spec a little so minor changes to the content
wording won't result in broken specs.