Inline form for identity confirmation task
Two extra locals and some conditional logic was added to the task form
partial to accommodate a couple of unique aspects to the identity
confirmation task. Things are about to get even more unique, as the
question itself, and the presence of the explanation is about to vary
depending on whether the claim has been identity-verified or not. To
stop the partial getting even more complicated let's inline the form on
the identity confirmation task and leave the form as the simple default
for the more straightforward tasks.