Add filebeat to web Docker image for log shipping
Filebeat is considered a more reliable way of sending log entries to
Logstash.

I attempted to run filebeat in a container instance and mount the logs
(from Azure file storage) to the container. However filebeat wasn't
picking up any changes to the file. On further investigation, they
recommend that you don't use network mounted files and instead install
filebeat on the same client as the logs[1].

Unfortunately filebeat don't have an official package available for
Alpine so we have to manually download and install it. To do this we
have to install `libc6-compat` to provide a libc6 compatibility layer
for the binary[2].

Because the filebeat command doesn't exit and outputs logs to std out we
have to use `&` to fork the process so the `bundle exec rails server`
command executes as well. It feels a little messy and we may want to run
it in the background/as a service.

[1] https://www.elastic.co/guide/en/beats/filebeat/master/filebeat-network-volumes.html
[2] https://discuss.elastic.co/t/filebeat-6-x-could-not-support-running-under-os-alpine/116195/11