Clarify requirements of rotating the certificates
It is necessary to be able to complete the GOV.UK Verify journey if
rotating the certificates on production to check the new keys and
certificates are working as expected.

It is mentioned futher down that a real account is needed but stating
the requirement up front may help decide who carries out and when the
rotation is done. Signing up to an identity provider isn't always
instant and may take several days.