Display decision on task view page for a claim
Displaying whether a claim has been approved or rejected will help
the service teams experience.

In my approach I created a claim_helper method format_claim_reference
to remove the logic for the header and decision out of the view pages.
The format_claim_reference method is being calculated within the _claim_summary
partial.
This seemed the appropriate place to do this as this is where the header
was being rendered.