Forbid amending student loan plan when no loan
In our current claim journey we only collect the claimant’s student loan
plan type if they tell us they’re still paying off their student loan.
This is because when we designed the claim journey, the only reason we
needed to know the plan type was so payroll could calculate the correct
student loan contributions to be deducted from the claim amount.

We recently had a case where a service operator amended a claim where
the claimant was no longer paying off their student loan, to set the
loan plan type to Plan 2.

Their reason for doing so was sensible – we also need to know the plan
type to be able to check the claim amount for TSLR claims. However, once
this claim was sent to Cantium, they saw the claim as having “no student
loan” and also having a plan type, which from their point of view
doesn’t make sense and which caused an error in their payroll system.

This problem would benefit from some service design thought – to find a
way to collect plan type even when the claimant has paid off their loan.
But until then, I just want to prevent the system from getting into this
state, so I’m adding a validation that prevents a service operator from
adding a plan type when the loan is already paid off. If a service
operator needs to collect this information – for checking the claim
amount – they should add a note to the claim.

I’ve suggested to the product owner that they add a card to their
backlog to explore this problem once they have development / service
design capacity.

dxw Zendesk tickets 11392, 11416.