Ignore undone rejections when anonymizing claims
A claim with an undone rejection should not be considered as rejected
and hence this rejection should not contribute to its eligibility to be
anonymized.

We recently accidentally scrubbed the personal data from some claims
which had been rejected and then subsequently approved, and only noticed
this when the payroll CSV file generation blew up because the payroll
gender was missing. We had to restore these claims' personal data from a
database backup.

https://dxw.zendesk.com/agent/tickets/11941