Remove banking_name when scrubbing personal data
It seems like we missed this one when we were choosing which fields to
scrub.

Our data retention policy [1] says that the only personal data we keep
are TRN and email address.

We already remove first_name, last_name, and all bank account numbers.
So I see no reason why we should be keeping banking_name.

This includes a data migration to fix the already-scrubbed claims.

[1] https://dfedigital.atlassian.net/wiki/spaces/TP/pages/1086193761/Data+handling+and+retention