Remove some extra detail from claim submitted email
The client requested that we remove the additional information about
phoning the claimant to verify their identity. The email is now the same
regardless of whether the claimant’s identity is already verified or
not. This is because they have had to “change (their) processes”.

This tweaks the original change made in e5a4543.