Reduce test environment alert sensitivity
Since we have set up StatusCake monitoring, we receive alerts almost every night.
This is probably due to the fact that test only has 1 app service instance so the
service is down if the app is restarted (which can be expected).

We want to keep the alert for the test environment but reduce the sensitivity.
The check interval is increased to 5 min and the trigger rate to 5 min.
This would allow downtime between 5 and 10 min.