Add request spec for PATCH /admin/claims/bulk_deallocate
Even though a feature spec is triggering this endpoint,
it may be worth adding a request spec to avoid slowing down the test suite
any further.

In general, especially if there is too much logic in a controller, it would be
advisable to test each endpoint in isolation in a request spec and keep a
coarse-grained approach to feature testing.

Testing only the currently changed endpoint.