Pass only the selected eligible subject to the ECP award calculator
In some cases, the claimant might have more than one subject or code;
The `AwardAmountCalculator` handles only one subject - which should be
the one selected by the user and if confirmed to be present in DQT,
either as a subject name or subject code.

This ensures only a valid subject "group" is used to validate if an
award amount can be calculated.