CAPT-1040 Make sure the "main" claim in `CurrentClaim` is correct
This commit addresses a bug that affects combined journeys (ECP/LUP at the moment)
during the post-eligibility phase.

Currently, the default "main" claim is always for ECP. Such claim receives all calls
redirected to the `Claim` model via `method_missing`. LUP receives updates only instead
(i.e. no validations, etc).

During the eligibility phase, no validations kick in, and both claims receive updates.
So far so good. During the claim phase, validations are run when calculating the next
page in the sequence. If any validation fails, instead of displaying the problem to the
user, they are forced to follow a "linear" sequence, which makes sense... because it
could mean they could have been somehow altering/skipping some answers.

But – it is a problem if the "main" claim in the `CurrentClaim` object is *not* the one
selected by the user at the end of the eligibility phase!

The bug reported in CAPT-1040 affects users who are eligible for LUP only (or chose LUP
instead of ECP despite the lower amount) and decide to go back to any question from the
final "Check your answers" page. In this specific scenario, they'd be forced to go
through all the remaining questions linearly, unnecessarily, simply because the "main"
claim in the `CurrentClaim` wrapper is always defaulting to the ECP claim, causing some
validations to fail.

With this commit, we are passing the policy selected at the end of the eligibility
phase to the `CurrentClaim` wrapper. Until then, or in the absence of a selected policy,
ECP can still be used as default.