Allow confirmation report batch uploading
This commit allows for multiple uploads for confirmation statements.

There is no restriction to the number of uploads as long as there are no duplicates.

When successful, an upload creates a payment confirmation with the scheduled
payment date in the CSV.

This commit also contains the logic to allow the download of a single ZIP file
containing all the batches for the payroll run.