Move rspec tag filter config
The RUN_* env variables can be evaluated in the rails_helper
because it's the one with the dotenv files loaded via Rails.

This commit also moves these env to the .env.test file,
defaulting them to `false`. A clearer comment explains why
they need to be enabled in a local environment.