Add index on QA-related columns
The number of rows that will enter the index is not going to be more
than 10% of all rows on the `claims` table, as we won't flag for QA
more that 10% of all claims. So it makes sense to create an index
and keep it partial. Also, we are interested in checking the `qa_completed_at`
value only when `qa_required = TRUE`. For any other queries fetching
the rest of the rows in the table (no less than 90%), there is no advantage
in having and using an index and a sequential scan will most likely be
forced anyway.