Auto-approve claims for which the SWC check returned NO DATA
This exception accounts for the fact that sometimes schools do not participate
in the school workforce census, and a NO DATA should not be considered as a
failure for the relevant task. In such cases, the task would normally be passed
manually, so it should not be a blocker to auto-approving claims.

This commit also removes some related but unused scopes.