Remove QTS Award from all categories except PG
Since we are inferring the correct category from the qualification name from DQT,
it's vital to make sure that 1) the list of qualifications is unique for each category,
2) in the presence of duplicates, the code cannot determine a category.

This commit enforces both points, and keeps QTS Award in the PG category only
as it's where most matches occur. In the future we may need to think of a different
way of deriving the correct category.