Reset eligibility attribute when dependent on a changed claim attribute
At the moment, you can reset only attributes that are dependent as long as they belong to the
same model; this commit introduces a more flexible approach where you can reset dependent
attributes on any associated model.

This is particularly needed for TSLR as a change in the NINO or DoB (stored on the claim),
must trigger a reset of the student loan plan type and repayment amount.