Retrieve and update claim student loan details when confirming personal details
This should work with the new TID journey as well. It will cost an extra update
to the current claim, but only at the end of the update controller action, so to
not interfere with other validations that may require form interaction.

If the retrieval and update of student loan details fails for any reason, we
should not prevent the user from continuing as the appropriate checks will still
run after submission. See next commits for further updates to the behaviour of
`ClaimStudentLoanDetailsUpdater` and validations.