Return student-loan-amount as the next slug when in sequence and updating personal details
For TSLR, the student-loan-amount slug needs to be the next in the sequence when
the applicant has answered to all the questions and decides to go back and change
personal details before submitting; if they change the NINO or DoB, we need to
display what information we have on their student loan repayments, which could be
different from what's been displayed before.

This is expected to work with TID as well.