Postpone retrieving SL details to later in the journey (TID only)
This commit's goal is to allow a policy eligibility criteria to be checked _before_ the student loan
(SL) criteria are.

This change is made necessary because on the TID route the personal details are fetched
automatically _before_ checking the eligibility criteria for a given policy: if we fetched the
SL details immediately as well, and the user were deemed ineligible (due to zero repayments),
we'd have to exit the journey immediately after they confirmed their TID details; in this
scenario the user wouldn't be able to tell if they could have been eligible based on subject
and qualification, for example.

This is an issue on TID only, as on the non-TID route the SL details are fetched _after_ the
eligibility is confirmed.

To solve the problem, we need to fetch the SL details in different moments based on wether
the user is claiming using the TID or non-TID route.

Just to make it slightly more complicated, there is an edge case on the TID route where the
SL details will be fetched in the same way they are on the non-TID route: when the personal
details that came through TID are incomplete or not valid, as in this case the user will be
presented with the screen to update their personal details after their eligibility is confirmed.