Extract NqtInAcademicYearAfterIttForm
Extracts a new form object and moves the logic associated with the
`nqt_in_academic_year_after_itt` step out of claims controller into the
form.
The view no longer needs to worry about the `eligibility` attributes
being nested, so it has been updated to remove the eligibility
namespacing on fields.

I'm not stoked about having to call url helpers in the form model for
the backlink path but was unsure how else to do this. Maybe we want to
consider introducing a new object `BacklinkPath` and pass the form
object to that? Something like
```
BacklinkPath.new(form).to_s #=> "/the-path/we/want"
```

Also I think if we're going to keep the permitting params logic in the
form object, we'll probably want to make it a bit more ergonomic to
define what params are permitted. We could add a class attribute to the
base form and then child forms could do something like
```
class SomeForm < Form
  permitted_params %i[some_param]
```
or we could just read off the attributes defined on the form and assume
any defined attribute is safe to permit as a paramerter.

We could also now remove the validation for
`:nqt_in_academic_year_after_itt` from
`Policies::EarlyCareerPayments::Eligibility` as it's no longer needed,
though I've left this in place for the time being, something for a
future commit.