Extracts form object for qualification step
We have to update the claim record slightly differently than usual to
ensure we trigger the reset dependent answers logic.

We likely want to pull out the reset dependent answers logic into each
form object, or potentially a different object that the forms call, to
simplify the responsibilities of the Claim model.