Remove callback
Removes the set_qualification_if_trainee_teacher call back from the
eligibility model.
Assuming the only place in we set `nqt_in_academic_year_after_itt` is
via the UI we can remove this callback from eligibility and let the form
object be responsible for setting this answer.

The NqtInAcademicYearAfterIttForm calls the QualificationForm to handle
setting the qualification if we're dealing with a trainee_teacher. It's
not ideal to couple these forms but I think it's better than duplicating
the dependent answer resetting logic. Perhaps we need to introduce an
object that's called after each form is saved which handles dealing with
resetting dependent answers but that is something for a future pr.