Extract form for teaching_subject_now
Extracts a form object for the teaching subject now step.

Depending on whether or not there's an eligible subject upon reaching
the `teaching-subject-now` step, we may redirect the subject to the
`eligbible-itt-subject` step rather than render the
`teaching-subject-now` form.
`eligible-itt-subject` can be reset if the `qualification` changes, and
`eligible-itt-subject` resets the `teaching-subject-now` answer.

For now we've preserved the existing behaviour of determining whether to
render the `teaching-subject-now` in the `ClaimsController#show` action.
This feels like the wrong place for this though. It makes more sense to
handle the redirect in the `ClaimsController#update` action or better
yet in the `PageSequence#next_slug` method, something to address in a
future commit.