Extract form for eligible_itt_subject
Currently the form object includes the IttSubjectHelper to access the
`subject_symbol`. Ideally we'd be able to remove this helper and have
the logic for determining what subjects are available handled in the
form. However, the `subject_symbols` method is used in various places in
the app and unpicking it isn't straightforward so we'll want to do that
in a separate commit.

Also of note, if the `subject_symbols` array is empty, which the helper
indicates is possible, then we render an empty page which is likely
something we don't want to do! Leaving this for now as that's the
existing behaviour and I need to find out what we should be doing.