Move methods into the form
Removes some of the view's dependencies on the 'subject_symbols' helper
moving this logic into the form object as, hopefully, a step towards
removing the subject_symbols helper method.
I looked into moving the logic of the `subject_symbols` method into the
form but pulling on that thread unravelled a lot of changes!