Move post sign in logic into oauth controller
When a teacher signs in with dfe Identity we want to store some details
on the current claim. We also need to reset the `details_check` on any
existing claims, to reshow the confirm details page, in case the teacher
has navigated back to the dfe identity page and signed in with a
different account.
Previously this was done in the claims controller flagged off the
presence of `:user_info` in the session (which was reset on the confirm
details show action and copied into the claim). We can remove this
concern from the claims controller, and avoid updating the claim in a
get request, by moving all this code into the omniauth call back
controller, this also allows us to store less data in the session.