Form object for sign-in-or-continue
By combining the sign_in_or_continue_form with the teacher_detail_form
we can avoid a lot of weirdness in the claims controller / auth callback, and
delete the teacher detail slug from the journey.

There's now two signin scenarios
User lands on sign in or continue form
1.) User selects "Skip dfe sign in" -> we PATCH to the claims controller
  update action with `logged_in_with_tid` `false` and they continue on
  their way.

2.) User selects "DfE sign-in" -> Users is redirected to DfE idenitiy ->
  User comes back to the omiauth_callbacks_controller with some
  additional info -> We redirect the user along with this additional
  info and the `logged_in_with_tid` flag set as `true` to the
  `sign-in-or-continue` form. As the sign-in-or-continue form sees the
  `logged_in_with_tid` flag as set it renders the confirm detail ui that
  was previously on the teacher-details page. The form can now PATCH to
  the ClaimsController#update action and the slug sequence will continue
  as normal. No need to remove slugs from the journey depending on
  whether or not the teacher as signed in with DfE identity.

This has the benefit of keeping all signin related logic in a single
form and doesn't require us to use the session to pass data between
controller.

We need to be a bit careful about logging parameters and also ensuring
any params from dfe sign in we redirect with are permitted - we don't
want to trigger an unremitted parameter error if dfe sign-in change the
payload. To ensure we're working with the expected parameters we slice
the ones we know about from the auth payload.

The SignInOrContinueForm uses a nested form TeacherIdUserInfoForm as we
need to write the attributes from DfEIdentity to the view, and using a
nested form lets us mimic the behaviour of
`accepts_nested_attributes_for` rather than having to store the JSON
`teacher_id_user_info` as some unusual structure in the view.

There's a potential refactoring we could do where we remove the
`DfeIdentity::UserInfo` class and place it's responsibilities in the
`TeacherIdUserInfoForm` however I avoided doing this to as it's a bit
out of scope.