Extract form object for teacher-detail
When the teacher comes back from dfe identity we present the details
from dfe identity back to the teacher for them to confirm.
If the teacher says the detail's don't match we blank them.
If the teacher says the details do match then we validate the details.
If the details validate we fetch the teacher's info from the DQT data
base.
If the details don't validate we leave them as is but set the details
flag to false and don't fetch the details from DQT.