Remove form deciding if slug should be set
If the teacher clicks to sign in with dfe identity, the
"sign-in-or-continue" form isn't submitted and so the list of visited
slugs isn't updated (as that is handled in the `ClaimsController#update`
action). This commit moves adding the slug to the session into the
omniauth callback controller, as if we've signed in with dfe identity
then we've visted the "sign-in-or-continue" page.