Add a form object for choosing the correct school
When using DfE Sign In, we present the school currently associated with
the user to check if it is still the correct school.

Currently, we retrieve the associated TPS school, store it in the
session so that we can display it in the view. Then on confirmation from
the user, this value gets persisted as the current school.

The form object allows us to extract this logic and it can define these
references, ensuring that the controller is simpler.

As far as I can tell, the session itself is unnecessary as the values
stored there are already persisted in the DB.