WIP - remove callback weirdness
By combinging the sign_in_or_continue_form with the teacher_detail_form
we can avoid a lot of weirdness in the claims controller / callback, and
delete the teacher detail slug from the journey.

There's now two signin scenarios
User lands on sign in or continue form
1.) User selects "Skip dfe sign in" -> we patch to the claims controller
  upudate action with `logged_in_with_tid` false and they continue on
  their way.

2.) User selects "DfE sign-in" -> Users is redirected to DfE idenitiy ->
  User comes back to the omiauth_callbacks_controller with some
  additional info -> We redirect the user along with this additional
  info and the `logged_in_with_tid` flag set as `true` to the
  `sign-in-or-continue` form. As the sign-in-or-continue form sees the
  `logged_in_with_tid` flag as set it renders the confirm detail ui that
  was previously on the teacher-details page. The form can now PATCH to
  the ClaimsController#update action and the slug sequence will continue
  as normal. No need to remove slugs from the journey depending on
  whether or not the teacher as signed in with DfE identity.

This has the benefit of keeping all signin related logic in a single
form and doesn't store any temporary user data in the session.

We need to be a bit careful about logging parameters and also ensuring
any params from dfe sign in we redirect with are permitted - we don't
want to trigger an unpermitted parameter error if dfe sign-in change the
payload. To ensure we're working with the expected parameters we slice
the ones we know about from the auth payload.