Add teacher_reference_number form
Adds a form object for teacher reference number.
When we upgrade rails and get support for Normalisation we should
replace overriding the `valid?` method with a normalisation call back.
The other thing I considered was adding ActiveModel::Callbacks to the
base form and defining a before validation callback there, might be
better not sure.

Have left some validations around trn on the claim model as it seems
like it can be set on the admin side. When we extract forms for the
admin ui we can remove these validations along with normalising the trn
from the claim model.