Use callback
On a pr review suggestion we've decided to use active model callbacks.
I don't think introducing callbacks in a form object should prove too
much of an issue as form objects are likely only to be used in one place
so it seems unlikely for there to be the usual callback surprises.