Extracts form mobile number step
We've moved sending the verification code into the form object. That seems to
be the right place for sending the code as it's a concern of updating the
mobile number. Ideally the form object would enqueue a background job to send
the code but that's for a future ticket!
Once the email verification step has been moved into a form we can remove the
`otp` method from the claims controller.
We don't call reset dependent attributes in the mobile form as we can just
reset the `mobile_verified` flag to nil in the same update.
There's a quirk where we check whether the response from Notify is `nil`. I'm
not sure how this could ever be `nil` but I've left the check in as this is
just a refactor pr.