Add a form object for postcode search
The postcode search step in the journey has some hardcoded control flow
in the controller.

As part of our drive to move to a form object approach, we can extract
the postcode search.

There's some things that aren't ideal about this change as it the
address selection steps are currently coupled to the postcode search.

The main issue here is that this changes the persistence of the postcode
value from the URLs and session to the claim itself.

The no-address-found screen appears to be an attempt to render the
invalid state of the postcode search, so I've made this explicit.

I have been hesitant to remove the no-address-found screen as I'm not
sure of the full implications there. However, I have included it in the
error state of the postcode search form.