Extracts form object for email
Following a similar pattern to the mobile number form, we move sending
the verification message into the form object as it's a concern of
providing/changing an email address.
We don't need to call reset dependent attributes as changing the email
address just nullifies `email_verified` which we can do inline in the
update method.
We're reusing the `sent_one_time_password_at` field between mobile and
email verification, it should be safe to do this as originally these two
verifications shared the same key in the session. We should investigate
what is involved in having a separate column for
`sent_mobile_one_time_password_at` and
`sent_email_one_time_password_at`. It's safe to split the
`sent_one_time_password_at` column later as the app is currently not
processing any claims in production.